+++ nodes/Apollo/utils.ts
@@ -1,0 +1,5 @@
+export const splitAndTrim = (str: string) =>
+	str
+		.split(';')
+		.map((s) => s.trim())
+		.filter((s) => s);